+++ src/components/feature-board-file-element/content/alternative-text/AlternativeText.vue
